# swagger/ywc-api-swagger.yaml
openapi: 3.0.3
info:
  title: YWC Financial Forms API
  description: |
    Comprehensive API for managing financial forms and client data for Your Wealth Coach (YWC).
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## User Roles
    - **ADMIN**: Full system access, can manage all users and data
    - **COACH**: Can manage assigned clients and their financial data
    - **CLIENT**: Can view and update their own financial information
    - **GUEST**: Limited read-only access to public endpoints
    
    ## User-Centric Architecture
    All financial data is now linked directly to users via `user_id`. When a client is created:
    1. A user account is created with role 'CLIENT'
    2. The client is assigned to a coach via `coach_id`
    3. All financial data (personal details, income, expenses, etc.) references the `user_id`
    
    ## Test Accounts
    - Admin: admin@ywc.com / admin123
    - Coach: coach@ywc.com / coach123  
    - Client: client@ywc.com / client123
    
  version: 2.0.0
  contact:
    name: YWC Development Team
    email: dev@yourwealthcoach.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.yourwealthcoach.com/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message about the operation
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message if success is false
      required:
        - success
        - message
...
    PersonalDetails:
      type: object
      properties:
        personal_id:
          type: string
          format: uuid
          description: Unique personal details identifier
        user_id:
          type: string
          format: uuid
          description: Associated user ID
        coach_user_id:
          type: string
          format: uuid
          nullable: true
          description: Assigned coach's user ID
        applicant_type:
          type: string
          enum: [PrimaryApplicant, SecondaryApplicant]
          description: Type of applicant
        salutation:
          type: string
          maxLength: 10
          description: Title (Mr., Mrs., Dr., etc.)
        first_name:
          type: string
          maxLength: 100
          description: First name
        last_name:
          type: string
          maxLength: 100
          description: Last name
        street:
          type: string
          maxLength: 255
          description: Street address
        house_number:
          type: string
          maxLength: 20
          description: House number
        postal_code:
          type: string
          maxLength: 20
          description: Postal/ZIP code
        city:
          type: string
          maxLength: 100
          description: City
        email:
          type: string
          format: email
          description: Email address
        phone:
          type: string
          maxLength: 20
          description: Phone number
        whatsapp:
          type: string
          maxLength: 20
          nullable: true
          description: WhatsApp number
        marital_status:
          type: string
          maxLength: 50
          description: Marital status
        birth_date:
          type: string
          format: date
          description: Date of birth
        birth_place:
          type: string
          maxLength: 100
          description: Place of birth
        nationality:
          type: string
          maxLength: 100
          description: Nationality
        residence_permit:
          type: string
          maxLength: 100
          nullable: true
          description: Residence permit information
        eu_citizen:
          type: boolean
          description: EU citizenship status
        tax_id:
          type: string
          maxLength: 100
          nullable: true
          description: Tax identification number
        iban:
          type: string
          maxLength: 50
          nullable: true
          description: International Bank Account Number
        housing:
          type: string
          maxLength: 100
          nullable: true
          description: Housing situation
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EmploymentDetails:
      type: object
      properties:
        employment_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        employment_type:
          type: string
          enum: [PrimaryEmployment, SecondaryEmployment]
        occupation:
          type: string
          maxLength: 200
        contract_type:
          type: string
          maxLength: 100
        contract_duration:
          type: string
          maxLength: 100
          nullable: true
        employer_name:
          type: string
          maxLength: 200
        employed_since:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IncomeDetails:
      type: object
      properties:
        income_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        gross_income:
          type: number
          format: decimal
          minimum: 0
          description: Gross monthly income
        net_income:
          type: number
          format: decimal
          minimum: 0
          description: Net monthly income
        tax_class:
          type: string
          maxLength: 10
          description: German tax class (1-6)
        tax_id:
          type: string
          maxLength: 100
          nullable: true
        number_of_salaries:
          type: integer
          minimum: 1
          maximum: 24
          default: 12
          description: Number of salary payments per year
        child_benefit:
          type: number
          format: decimal
          minimum: 0
          default: 0
        other_income:
          type: number
          format: decimal
          minimum: 0
          default: 0
        income_trade_business:
          type: number
          format: decimal
          minimum: 0
          default: 0
        income_self_employed_work:
          type: number
          format: decimal
          minimum: 0
          default: 0
        income_side_job:
          type: number
          format: decimal
          minimum: 0
          default: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExpensesDetails:
      type: object
      properties:
        expenses_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        cold_rent:
          type: number
          format: decimal
          minimum: 0
          default: 0
        electricity:
          type: number
          format: decimal
          minimum: 0
          default: 0
        living_expenses:
          type: number
          format: decimal
          minimum: 0
          default: 0
        gas:
          type: number
          format: decimal
          minimum: 0
          default: 0
        telecommunication:
          type: number
          format: decimal
          minimum: 0
          default: 0
        account_maintenance_fee:
          type: number
          format: decimal
          minimum: 0
          default: 0
        alimony:
          type: number
          format: decimal
          minimum: 0
          default: 0
        subscriptions:
          type: number
          format: decimal
          minimum: 0
          default: 0
        other_expenses:
          type: number
          format: decimal
          minimum: 0
          default: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Asset:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        real_estate:
          type: number
          format: decimal
          minimum: 0
          default: 0
        securities:
          type: number
          format: decimal
          minimum: 0
          default: 0
        bank_deposits:
          type: number
          format: decimal
          minimum: 0
          default: 0
        building_savings:
          type: number
          format: decimal
          minimum: 0
          default: 0
        insurance_values:
          type: number
          format: decimal
          minimum: 0
          default: 0
        other_assets:
          type: number
          format: decimal
          minimum: 0
          default: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Liability:
      type: object
      properties:
        liability_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        loan_type:
          type: string
          enum: [PersonalLoan, HomeLoan, CarLoan, BusinessLoan, EducationLoan, OtherLoan]
        loan_bank:
          type: string
          maxLength: 200
          nullable: true
        loan_amount:
          type: number
          format: decimal
          minimum: 0
          nullable: true
        loan_monthly_rate:
          type: number
          format: decimal
          minimum: 0
          nullable: true
        loan_interest:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
          nullable: true
          description: Interest rate as percentage
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Operation failed"
        error:
          type: string
          example: "Detailed error message"
paths:
...
  /personal-details:
    post:
      tags:
        - Personal Details
      summary: Create personal details
      description: Create personal details for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicant_type:
                  type: string
                  enum: [PrimaryApplicant, SecondaryApplicant]
                salutation:
                  type: string
                  maxLength: 10
                first_name:
                  type: string
                  maxLength: 100
                last_name:
                  type: string
                  maxLength: 100
                street:
                  type: string
                  maxLength: 255
                house_number:
                  type: string
                  maxLength: 20
                postal_code:
                  type: string
                  maxLength: 20
                city:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  maxLength: 20
                whatsapp:
                  type: string
                  maxLength: 20
                marital_status:
                  type: string
                  maxLength: 50
                birth_date:
                  type: string
                  format: date
                birth_place:
                  type: string
                  maxLength: 100
                nationality:
                  type: string
                  maxLength: 100
                residence_permit:
                  type: string
                  maxLength: 100
                eu_citizen:
                  type: boolean
                tax_id:
                  type: string
                  maxLength: 100
                iban:
                  type: string
                  maxLength: 50
                housing:
                  type: string
                  maxLength: 100
              required:
                - applicant_type
                - salutation
                - first_name
                - last_name
                - street
                - house_number
                - postal_code
                - city
                - email
                - phone
                - marital_status
                - birth_date
                - birth_place
                - nationality
                - eu_citizen
      responses:
        '201':
          description: Personal details created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PersonalDetails'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Personal Details
      summary: Get all personal details
      description: Retrieve all personal details with optional filtering
      parameters:
        - name: applicant_type
          in: query
          schema:
            type: string
            enum: [PrimaryApplicant, SecondaryApplicant]
          description: Filter by applicant type
        - name: marital_status
          in: query
          schema:
            type: string
          description: Filter by marital status
      responses:
        '200':
          description: Personal details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PersonalDetails'
...
  /employment:
    post:
      tags:
        - Employment
      summary: Create employment details
      description: Create new employment details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employment_type:
                  type: string
                  enum: [PrimaryEmployment, SecondaryEmployment]
                occupation:
                  type: string
                  maxLength: 200
                contract_type:
                  type: string
                  maxLength: 100
                contract_duration:
                  type: string
                  maxLength: 100
                employer_name:
                  type: string
                  maxLength: 200
                employed_since:
                  type: string
                  format: date
              required:
                - employment_type
                - occupation
                - contract_type
                - employer_name
                - employed_since
      responses:
        '201':
          description: Employment details created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmploymentDetails'
...
  /income:
    post:
      tags:
        - Income
      summary: Create income details
      description: Create new income details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gross_income:
                  type: number
                  format: decimal
                  minimum: 0
                net_income:
                  type: number
                  format: decimal
                  minimum: 0
                tax_class:
                  type: string
                  maxLength: 10
                tax_id:
                  type: string
                  maxLength: 100
                number_of_salaries:
                  type: integer
                  minimum: 1
                  maximum: 24
                  default: 12
                child_benefit:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                other_income:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                income_trade_business:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                income_self_employed_work:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                income_side_job:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
              required:
                - gross_income
                - net_income
                - tax_class
      responses:
        '201':
          description: Income details created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IncomeDetails'

    get:
      tags:
        - Income
      summary: Get all income details
      description: Retrieve all income details
      responses:
        '200':
          description: Income details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/IncomeDetails'
...
  /expenses:
    post:
      tags:
        - Expenses
      summary: Create expense details
      description: Create new expense details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cold_rent:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                electricity:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                living_expenses:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                gas:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                telecommunication:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                account_maintenance_fee:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                alimony:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                subscriptions:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                other_expenses:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
      responses:
        '201':
          description: Expense details created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExpensesDetails'

    get:
      tags:
        - Expenses
      summary: Get all expense details
      description: Retrieve all expense details
      responses:
        '200':
          description: Expense details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExpensesDetails'
...
  /assets:
    post:
      tags:
        - Assets
      summary: Create asset record
      description: Create new asset record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                real_estate:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                securities:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                bank_deposits:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                building_savings:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                insurance_values:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
                other_assets:
                  type: number
                  format: decimal
                  minimum: 0
                  default: 0
      responses:
        '201':
          description: Asset record created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Asset'

    get:
      tags:
        - Assets
      summary: Get all asset records
      description: Retrieve all asset records
      responses:
        '200':
          description: Asset records retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Asset'
...
  /liabilities:
    post:
      tags:
        - Liabilities
      summary: Create liability record
      description: Create new liability record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                loan_type:
                  type: string
                  enum: [PersonalLoan, HomeLoan, CarLoan, BusinessLoan, EducationLoan, OtherLoan]
                loan_bank:
                  type: string
                  maxLength: 200
                loan_amount:
                  type: number
                  format: decimal
                  minimum: 0
                loan_monthly_rate:
                  type: number
                  format: decimal
                  minimum: 0
                loan_interest:
                  type: number
                  format: decimal
                  minimum: 0
                  maximum: 100
              required:
                - loan_type
      responses:
        '201':
          description: Liability record created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Liability'

    get:
      tags:
        - Liabilities
      summary: Get all liability records
      description: Retrieve all liability records
      responses:
        '200':
          description: Liability records retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Liability'
...